테스트 전용 모델 패키지를 fulltext.fixture 하위로 옮겼다
