+++ src/test/java/fulltext/fixture/model/InvalidClassAnnotationModel.java
@@ -1,41 +1,50 @@
-package fulltext.model;
+package fulltext.fixture.model;
 
 import fulltext.annotation.Field;
-import fulltext.annotation.FullText;
 import fulltext.enums.Charset;
+import fulltext.annotation.FullText;
 import fulltext.enums.PadCharacter;
 import java.time.LocalDate;
 import java.util.Objects;
 
 @FullText(
-    length = 300,
+    length = 301,
     encoding = Charset.UTF_8, // 명시하지 않을 경우 기본값은 UTF-8입니다.
     padChar = PadCharacter.SPACE // 명시하지 않을 경우 기본값은 공백문자(" ")입니다.
 )
-public class NoFieldAnnotationModel {
+public class InvalidClassAnnotationModel {
 
+    @Field(length = 1)
     private String headerType;
 
+    @Field(length = 8)
     private LocalDate createAt; // yyyyMMdd
 
+    @Field(length = 91)
     private String headerPadding;
 
+    @Field(length = 1)
     private String dataType;
 
+    @Field(length = 10)
     private String name;
 
+    @Field(length = 3)
     private int age;
 
+    @Field(length = 86)
     private String dataPadding;
 
+    @Field(length = 1)
     private String trailerType;
 
+    @Field(length = 99)
     private String trailerPadding;
 
-    private NoFieldAnnotationModel() {
+    private InvalidClassAnnotationModel() {
     }
 
-    private NoFieldAnnotationModel(final String headerType, final LocalDate createAt, final String headerPadding, final String dataType,
+    private InvalidClassAnnotationModel(final String headerType, final LocalDate createAt, final String headerPadding, final String dataType,
         final String name, final int age, final String dataPadding, final String trailerType, final String trailerPadding) {
         this.headerType = headerType;
         this.createAt = createAt;
@@ -48,8 +57,8 @@ public class NoFieldAnnotationModel {
         this.trailerPadding = trailerPadding;
     }
 
-    public static ValidModelBuilder builder() {
-        return new ValidModelBuilder();
+    public static InvalidClassAnnotationModelBuilder builder() {
+        return new InvalidClassAnnotationModelBuilder();
     }
 
     @Override
@@ -75,7 +84,7 @@ public class NoFieldAnnotationModel {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        final NoFieldAnnotationModel validModel = (NoFieldAnnotationModel) o;
+        final InvalidClassAnnotationModel validModel = (InvalidClassAnnotationModel) o;
         return age == validModel.age && Objects.equals(headerType, validModel.headerType) && Objects.equals(createAt, validModel.createAt) && Objects.equals(headerPadding, validModel.headerPadding) && Objects.equals(
             dataType, validModel.dataType) && Objects.equals(name, validModel.name) && Objects.equals(dataPadding, validModel.dataPadding) && Objects.equals(trailerType, validModel.trailerType) && Objects.equals(
             trailerPadding, validModel.trailerPadding);
@@ -86,7 +95,7 @@ public class NoFieldAnnotationModel {
         return Objects.hash(headerType, createAt, headerPadding, dataType, name, age, dataPadding, trailerType, trailerPadding);
     }
 
-    public static class ValidModelBuilder {
+    public static class InvalidClassAnnotationModelBuilder {
 
         private String headerType;
         private LocalDate createAt;
@@ -98,53 +107,53 @@ public class NoFieldAnnotationModel {
         private String trailerType;
         private String trailerPadding;
 
-        public ValidModelBuilder headerType(final String headerType) {
+        public InvalidClassAnnotationModelBuilder headerType(final String headerType) {
             this.headerType = headerType;
             return this;
         }
 
-        public ValidModelBuilder createAt(final LocalDate createAt) {
+        public InvalidClassAnnotationModelBuilder createAt(final LocalDate createAt) {
             this.createAt = createAt;
             return this;
         }
 
-        public ValidModelBuilder headerPadding(final String headerPadding) {
+        public InvalidClassAnnotationModelBuilder headerPadding(final String headerPadding) {
             this.headerPadding = headerPadding;
             return this;
         }
 
-        public ValidModelBuilder dataType(final String dataType) {
+        public InvalidClassAnnotationModelBuilder dataType(final String dataType) {
             this.dataType = dataType;
             return this;
         }
 
-        public ValidModelBuilder name(final String name) {
+        public InvalidClassAnnotationModelBuilder name(final String name) {
             this.name = name;
             return this;
         }
 
-        public ValidModelBuilder age(final int age) {
+        public InvalidClassAnnotationModelBuilder age(final int age) {
             this.age = age;
             return this;
         }
 
-        public ValidModelBuilder dataPadding(final String dataPadding) {
+        public InvalidClassAnnotationModelBuilder dataPadding(final String dataPadding) {
             this.dataPadding = dataPadding;
             return this;
         }
 
-        public ValidModelBuilder trailerType(final String trailerType) {
+        public InvalidClassAnnotationModelBuilder trailerType(final String trailerType) {
             this.trailerType = trailerType;
             return this;
         }
 
-        public ValidModelBuilder trailerPadding(final String trailerPadding) {
+        public InvalidClassAnnotationModelBuilder trailerPadding(final String trailerPadding) {
             this.trailerPadding = trailerPadding;
             return this;
         }
 
-        public NoFieldAnnotationModel build() {
-            return new NoFieldAnnotationModel(headerType, createAt, headerPadding, dataType, name, age, dataPadding, trailerType, trailerPadding);
+        public InvalidClassAnnotationModel build() {
+            return new InvalidClassAnnotationModel(headerType, createAt, headerPadding, dataType, name, age, dataPadding, trailerType, trailerPadding);
         }
 
     }
