+++ src/test/java/fulltext/fixture/model/ValidModel.java
@@ -1,20 +1,18 @@
-package fulltext.model;
+package fulltext.fixture.model;
 
 import fulltext.annotation.Field;
-import fulltext.annotation.FullText;
 import fulltext.enums.Charset;
+import fulltext.annotation.FullText;
 import fulltext.enums.PadCharacter;
-import fulltext.enums.PadPosition;
 import java.time.LocalDate;
 import java.util.Objects;
 
 @FullText(
     length = 300,
     encoding = Charset.UTF_8, // 명시하지 않을 경우 기본값은 UTF-8입니다.
-    padChar = PadCharacter.SPACE, // 명시하지 않을 경우 기본값은 공백문자(" ")입니다.
-    padPosition = PadPosition.LEFT_PAD // 명시하지 않을 경우 기본적으로 왼쪽에 패딩문자를 채워넣습니다.
+    padChar = PadCharacter.SPACE // 명시하지 않을 경우 기본값은 공백문자(" ")입니다.
 )
-public class ValidOptionModel {
+public class ValidModel {
 
     @Field(length = 1)
     private String headerType;
@@ -28,10 +26,10 @@ public class ValidOptionModel {
     @Field(length = 1)
     private String dataType;
 
-    @Field(length = 10, padPosition = PadPosition.RIGHT_PAD) // @Field의 속성이 @FullText보다 우선됩니다.
+    @Field(length = 10)
     private String name;
 
-    @Field(length = 3, padChar = PadCharacter.ZERO) // @Field의 속성이 @FullText보다 우선됩니다.
+    @Field(length = 3)
     private int age;
 
     @Field(length = 86)
@@ -43,10 +41,10 @@ public class ValidOptionModel {
     @Field(length = 99)
     private String trailerPadding;
 
-    private ValidOptionModel() {
+    private ValidModel() {
     }
 
-    private ValidOptionModel(final String headerType, final LocalDate createAt, final String headerPadding, final String dataType,
+    private ValidModel(final String headerType, final LocalDate createAt, final String headerPadding, final String dataType,
         final String name, final int age, final String dataPadding, final String trailerType, final String trailerPadding) {
         this.headerType = headerType;
         this.createAt = createAt;
@@ -59,8 +57,8 @@ public class ValidOptionModel {
         this.trailerPadding = trailerPadding;
     }
 
-    public static ValidOptionModelBuilder builder() {
-        return new ValidOptionModelBuilder();
+    public static ValidModelBuilder builder() {
+        return new ValidModelBuilder();
     }
 
     @Override
@@ -86,7 +84,7 @@ public class ValidOptionModel {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        final ValidOptionModel validModel = (ValidOptionModel) o;
+        final ValidModel validModel = (ValidModel) o;
         return age == validModel.age && Objects.equals(headerType, validModel.headerType) && Objects.equals(createAt, validModel.createAt) && Objects.equals(headerPadding, validModel.headerPadding) && Objects.equals(
             dataType, validModel.dataType) && Objects.equals(name, validModel.name) && Objects.equals(dataPadding, validModel.dataPadding) && Objects.equals(trailerType, validModel.trailerType) && Objects.equals(
             trailerPadding, validModel.trailerPadding);
@@ -97,7 +95,7 @@ public class ValidOptionModel {
         return Objects.hash(headerType, createAt, headerPadding, dataType, name, age, dataPadding, trailerType, trailerPadding);
     }
 
-    public static class ValidOptionModelBuilder {
+    public static class ValidModelBuilder {
 
         private String headerType;
         private LocalDate createAt;
@@ -109,53 +107,53 @@ public class ValidOptionModel {
         private String trailerType;
         private String trailerPadding;
 
-        public ValidOptionModelBuilder headerType(final String headerType) {
+        public ValidModelBuilder headerType(final String headerType) {
             this.headerType = headerType;
             return this;
         }
 
-        public ValidOptionModelBuilder createAt(final LocalDate createAt) {
+        public ValidModelBuilder createAt(final LocalDate createAt) {
             this.createAt = createAt;
             return this;
         }
 
-        public ValidOptionModelBuilder headerPadding(final String headerPadding) {
+        public ValidModelBuilder headerPadding(final String headerPadding) {
             this.headerPadding = headerPadding;
             return this;
         }
 
-        public ValidOptionModelBuilder dataType(final String dataType) {
+        public ValidModelBuilder dataType(final String dataType) {
             this.dataType = dataType;
             return this;
         }
 
-        public ValidOptionModelBuilder name(final String name) {
+        public ValidModelBuilder name(final String name) {
             this.name = name;
             return this;
         }
 
-        public ValidOptionModelBuilder age(final int age) {
+        public ValidModelBuilder age(final int age) {
             this.age = age;
             return this;
         }
 
-        public ValidOptionModelBuilder dataPadding(final String dataPadding) {
+        public ValidModelBuilder dataPadding(final String dataPadding) {
             this.dataPadding = dataPadding;
             return this;
         }
 
-        public ValidOptionModelBuilder trailerType(final String trailerType) {
+        public ValidModelBuilder trailerType(final String trailerType) {
             this.trailerType = trailerType;
             return this;
         }
 
-        public ValidOptionModelBuilder trailerPadding(final String trailerPadding) {
+        public ValidModelBuilder trailerPadding(final String trailerPadding) {
             this.trailerPadding = trailerPadding;
             return this;
         }
 
-        public ValidOptionModel build() {
-            return new ValidOptionModel(headerType, createAt, headerPadding, dataType, name, age, dataPadding, trailerType, trailerPadding);
+        public ValidModel build() {
+            return new ValidModel(headerType, createAt, headerPadding, dataType, name, age, dataPadding, trailerType, trailerPadding);
         }
 
     }
